requirements.txt
aws-cdk-lib==2.143.0
constructs>=10.0.0,<11.0.0
requests
aws_cdk.aws_lambda_python_alpha
aws-lambda-powertools[all]
cdklabs.generative_ai_cdk_constructs
